lookupCSG function
Will, I dont know what you were doing. It seemed like you were unsure of
whether to write a lookup function for just CSG's, or a master lookup
which can lookup anything. Regardless, with the change to chained
buckets, I figured it would be easier to just rewrite the function. This
function only looks up CSG's, and it probably has a lot of bugs. More to
follow.